requirements.txt
pandas
google-cloud-core
google-cloud-storage
google-cloud-automl
tensorflow==1.13.1
